+++ DataTests/Http/HttpClientSpy.swift
@@ -12,7 +12,7 @@ import Data
 class HttpClientSpy : HttpPostClient {
     var urls = [URL]()
     var data: Data?
-    var completion: ((Result<Data, HttpError>) -> Void)?
+    var completion: ((Result<Data?, HttpError>) -> Void)?
     
     func post(to url: URL, with data: Data?, completion: @escaping (Result<Data?, HttpError>) -> Void) {
         self.urls.append(url)
