--- Infra/AlamofireAdapter.swift
@@ -1,9 +1,0 @@
-//
-//  AlamofireAdapter.swift
-//  Infra
-//
-//  Created by Israel Ermel on 18/08/20.
-//  Copyright © 2020 Israel Ermel. All rights reserved.
-//
-
-import Foundation
