+++ InfraTests/Services/AlamofireAdapterTests.swift
@@ -9,44 +9,7 @@
 import XCTest
 import Alamofire
 import Data
-
-class AlamofireAdapter: HttpPostClient {
-    
-    private let session: Session
-    
-    init(session: Session = .default) {
-        self.session = session
-    }
-    
-    func post(to url: URL, with data: Data?, completion: @escaping (Result<Data?, HttpError>) -> Void) {
-        self.session.request(url, method: .post, parameters: data?.toJson(), encoding: JSONEncoding.default).responseData {
-            dataReponse in
-            
-            guard let statusCode = dataReponse.response?.statusCode else { return completion(.failure(.noConnectivity)) }
-            
-            switch dataReponse.result {
-            case .failure: completion(.failure(.noConnectivity))
-            case .success(let data):
-                switch statusCode {
-                case 204:
-                    completion(.success(nil))
-                case 200...299:
-                    completion(.success(data))
-                case 401:
-                    completion(.failure(.unauthorized))
-                case 403:
-                    completion(.failure(.forbidden))
-                case 400...499:
-                    completion(.failure(.badRequest))
-                case 500...599:
-                    completion(.failure(.serverError))
-                default:
-                    completion(.failure(.noConnectivity))
-                }
-            }
-        }
-    }
-}
+import Infra
 
 class AlamofireAdapterTests: XCTestCase {
     
@@ -72,11 +35,18 @@ class AlamofireAdapterTests: XCTestCase {
     
     func test_post_should_complete_with_error_on_all_invalid_cases(){
         expectResult(.failure(.noConnectivity), when: (data: nil, response: makeHttpResponse(), error: nil))
+        
         expectResult(.failure(.noConnectivity), when: (data: makeValidData(), response: nil, error: makeError()))
+        
         expectResult(.failure(.noConnectivity), when: (data: makeValidData(), response: nil, error: nil))
+        
         expectResult(.failure(.noConnectivity), when: (data: nil, response: makeHttpResponse(), error: makeError()))
+        
         expectResult(.failure(.noConnectivity), when: (data: nil, response: makeHttpResponse(), error: nil))
+        
         expectResult(.failure(.noConnectivity), when: (data: nil, response: nil, error: nil))
+     
+        expectResult(.failure(.noConnectivity), when: (data: makeValidData(), response: makeHttpResponse(statusCode: 300), error: nil))
         
     }
     
@@ -154,48 +124,4 @@ extension AlamofireAdapterTests {
     
 }
 
-class UrlProtocolStub: URLProtocol {
-    static var emit: ((URLRequest) -> Void)?
-    
-    static var data: Data?
-    static var response: HTTPURLResponse?
-    static var error: Error?
-    
-    static func simulate(data: Data?, response: HTTPURLResponse?, error:Error?) {
-        UrlProtocolStub.data = data
-        UrlProtocolStub.response = response
-        UrlProtocolStub.error = error
-    }
-    
-    static func observeRequest(completion: @escaping (URLRequest) -> Void) {
-        UrlProtocolStub.emit = completion
-    }
-    
-    override class func canInit(with request: URLRequest) -> Bool {
-        return true
-    }
-    
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        request
-    }
-    
-    override func startLoading() {
-        UrlProtocolStub.emit?(request)
-        
-        if let data = UrlProtocolStub.data {
-            client?.urlProtocol(self, didLoad: data)
-        }
-        
-        if let response = UrlProtocolStub.response {
-            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-        }
-        
-        if let error = UrlProtocolStub.error {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-        
-        client?.urlProtocolDidFinishLoading(self)
-    }
-    
-    override func stopLoading() {}
-}
+
