+++ InfraTests/Services/UrlProtocolStub.swift
@@ -1,0 +1,55 @@
+//
+//  UrlProtocolStub.swift
+//  InfraTests
+//
+//  Created by Israel Ermel on 26/08/20.
+//  Copyright © 2020 Israel Ermel. All rights reserved.
+//
+
+import Foundation
+
+class UrlProtocolStub: URLProtocol {
+    static var emit: ((URLRequest) -> Void)?
+    
+    static var data: Data?
+    static var response: HTTPURLResponse?
+    static var error: Error?
+    
+    static func simulate(data: Data?, response: HTTPURLResponse?, error:Error?) {
+        UrlProtocolStub.data = data
+        UrlProtocolStub.response = response
+        UrlProtocolStub.error = error
+    }
+    
+    static func observeRequest(completion: @escaping (URLRequest) -> Void) {
+        UrlProtocolStub.emit = completion
+    }
+    
+    override class func canInit(with request: URLRequest) -> Bool {
+        return true
+    }
+    
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        request
+    }
+    
+    override func startLoading() {
+        UrlProtocolStub.emit?(request)
+        
+        if let data = UrlProtocolStub.data {
+            client?.urlProtocol(self, didLoad: data)
+        }
+        
+        if let response = UrlProtocolStub.response {
+            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+        }
+        
+        if let error = UrlProtocolStub.error {
+            client?.urlProtocol(self, didFailWithError: error)
+        }
+        
+        client?.urlProtocolDidFinishLoading(self)
+    }
+    
+    override func stopLoading() {}
+}
